[백] jwt 토큰 이용해서 현재 로그인한 멤버의 정보 DB에서 조회하기 성공, 그에 따른 로그인한 사용자의 기부내역만 보여주기 구현 이전 기능들 중 멤버에 관한 기능 추가 개발중

+++ src/main/java/metaint/replanet/rest/pay/entity/Member.java
@@ -17,7 +17,7 @@ public class Member {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "member_code")
-    private int memberCode;
+    private Long memberCode;
     @Column(name = "member_id")
     private String memberId;
     @Column(name = "member_name")
